feat: added implementation for Company class to return specific company, also added company not found exception class

+++ src/main/java/com/thoughtworks/springbootemployee/controller/Company.java
@@ -1,14 +1,14 @@
 package com.thoughtworks.springbootemployee.controller;public class Company {
-    private final int companyId;
+    private final Long companyId;
     private final String companyName;
 
-    public Company(int companyId, String companyName) {
+    public Company(Long companyId, String companyName) {
 
         this.companyId = companyId;
         this.companyName = companyName;
     }
 
-    public int getCompanyId() {
+    public Long getCompanyId() {
         return companyId;
     }
